refactor: Restruct package architecture
