+++ scrapper/src/main/kotlin/ru/bytebosses/scrapper/api/chat/exception/ChatAlreadyExistException.kt
@@ -1,3 +1,3 @@
-package ru.bytebosses.scrapper.exception
+package ru.bytebosses.scrapper.api.chat.exception
 
 class ChatAlreadyExistException(val id: Long) : RuntimeException("Chat $id already exists")
