+++ scrapper/src/main/kotlin/ru/bytebosses/scrapper/api/chat/exception/ChatIsNotExistException.kt
@@ -1,3 +1,3 @@
-package ru.bytebosses.scrapper.exception
+package ru.bytebosses.scrapper.api.chat.exception
 
 class ChatIsNotExistException(chatId: Long) : RuntimeException("Chat $chatId is not exists")
