+++ scrapper/src/main/kotlin/ru/bytebosses/scrapper/sender/UpdateSender.kt
@@ -1,6 +1,6 @@
 package ru.bytebosses.scrapper.sender
 
-import ru.bytebosses.scrapper.dto.request.bot.LinkUpdate
+import ru.bytebosses.scrapper.api.common.dto.request.bot.LinkUpdate
 
 interface UpdateSender {
     fun sendUpdate(update: LinkUpdate)
