+++ scrapper/src/main/kotlin/ru/bytebosses/scrapper/service/LinkService.kt
@@ -1,6 +1,6 @@
 package ru.bytebosses.scrapper.service
 
-import ru.bytebosses.scrapper.dto.response.LinkResponse
+import ru.bytebosses.scrapper.api.link.dto.LinkResponse
 import ru.bytebosses.scrapper.model.Link
 import java.net.URI
 import kotlin.time.Duration
